+++ LongevityWorldCup.Website/Tools/ThreadsMessageBuilder.cs
@@ -57,7 +57,7 @@ public static class ThreadsMessageBuilder
             if (!EventHelpers.TryExtractBiologicalAgeImprovement(rawText, out var clock, out var fromAge, out var toAge)) return "";
             var athlete = slugToName(improvementSlug);
             var clockContext = string.Equals(clock, "bortz", StringComparison.OrdinalIgnoreCase) ? "bortz" : "pheno";
-            return RejectIfTooLong(BuildBiologicalAgeImprovementLine(athlete, clock, fromAge, toAge, SocialPostLinks.AthleteUrl(improvementSlug, clockContext)));
+            return RejectIfTooLong(SocialPostText.BuildBiologicalAgeImprovementLine(athlete, clock, fromAge, toAge, SocialPostLinks.AthleteUrl(improvementSlug, clockContext)));
         }
 
         if (type == EventType.CrowdAgeTop10Change)
@@ -67,7 +67,7 @@ public static class ThreadsMessageBuilder
             EventHelpers.TryExtractPrev(rawText, out var prevSlug);
             var athlete = slugToName(crowdSlug);
             var prev = !string.IsNullOrWhiteSpace(prevSlug) ? slugToName(prevSlug) : null;
-            return RejectIfTooLong(BuildCrowdAgeTop10Line(athlete, crowdPlace, previousPlace, prev, crowdAge, crowdCount, getChronoAgeForSlug?.Invoke(crowdSlug), SocialPostLinks.AthleteUrl(crowdSlug, "crowd")));
+            return RejectIfTooLong(SocialPostText.BuildCrowdAgeTop10Line(athlete, crowdPlace, previousPlace, prev, crowdAge, crowdCount, getChronoAgeForSlug?.Invoke(crowdSlug), SocialPostLinks.AthleteUrl(crowdSlug, "crowd")));
         }
 
         if (type == EventType.AgeImprovementTop10Change)
@@ -78,7 +78,7 @@ public static class ThreadsMessageBuilder
             var athlete = slugToName(improvementLeaderboardSlug);
             var prev = !string.IsNullOrWhiteSpace(prevSlug) ? slugToName(prevSlug) : null;
             var leagueCtxSlug = string.Equals(improvementClock, "bortz", StringComparison.OrdinalIgnoreCase) ? "bortz-improvement" : "improvement";
-            return RejectIfTooLong(BuildAgeImprovementTop10Line(athlete, improvementClock, improvementPlace, previousPlace, prev, improvement, SocialPostLinks.AthleteUrl(improvementLeaderboardSlug, leagueCtxSlug)));
+            return RejectIfTooLong(SocialPostText.BuildAgeImprovementTop10Line(athlete, improvementClock, improvementPlace, previousPlace, prev, improvement, SocialPostLinks.AthleteUrl(improvementLeaderboardSlug, leagueCtxSlug)));
         }
 
         if (type != EventType.BadgeAward) return "";
@@ -438,115 +438,6 @@ public static class ThreadsMessageBuilder
     private static string BuildBecameProLine(string athleteName, string athleteUrl)
     {
         return $"{athleteName} went Pro.\n\nBortz Age results now place them in the Pro track.\n\n{athleteUrl}";
-    }
-
-    private static string BuildBiologicalAgeImprovementLine(
-        string athleteName,
-        string clock,
-        double fromAge,
-        double toAge,
-        string athleteUrl)
-    {
-        var clockLabel = string.Equals(clock, "bortz", StringComparison.OrdinalIgnoreCase)
-            ? "Bortz Age"
-            : "pheno age";
-        var fromText = fromAge.ToString("0.##", CultureInfo.InvariantCulture);
-        var toText = toAge.ToString("0.##", CultureInfo.InvariantCulture);
-        return $"{athleteName} improved their {clockLabel} from {fromText} to {toText} years.\n\n{athleteUrl}";
-    }
-
-    private static string FormatSignedYears(double years)
-    {
-        var text = years.ToString("0.#", CultureInfo.InvariantCulture);
-        return years > 0 ? $"+{text}" : text;
-    }
-
-    private static string BuildCrowdAgeTop10Line(
-        string athleteName,
-        int place,
-        int? previousPlace,
-        string? previousAthlete,
-        double crowdAge,
-        int crowdCount,
-        double? chronologicalAge,
-        string athleteUrl)
-    {
-        var crowdAgeText = crowdAge.ToString("0.#", CultureInfo.InvariantCulture);
-        var countText = crowdCount.ToString("N0", CultureInfo.InvariantCulture);
-        var movement = BuildCrowdAgeMovement(place, previousPlace);
-        var signal = BuildCrowdAgeSignal(crowdAge, chronologicalAge);
-        var metricLine = !string.IsNullOrWhiteSpace(signal)
-            ? $"{athleteName}'s Crowd Age is {crowdAgeText}, {signal}."
-            : $"{athleteName}'s Crowd Age is {crowdAgeText}.";
-
-        return $"{athleteName} {movement} in Crowd Age with {countText} guesses.\n{metricLine}\n\n{athleteUrl}";
-    }
-
-    private static string BuildCrowdAgeMovement(int place, int? previousPlace)
-    {
-        var placeText = CrowdOrdinal(place);
-        return previousPlace.HasValue
-            ? previousPlace.Value > place
-                ? $"climbed from {CrowdOrdinal(previousPlace.Value)} to {placeText}"
-                : $"moved from {CrowdOrdinal(previousPlace.Value)} to {placeText}"
-            : $"just entered the top 10 at {placeText}";
-    }
-
-    private static string? BuildCrowdAgeSignal(double crowdAge, double? chronologicalAge)
-    {
-        if (!chronologicalAge.HasValue || !double.IsFinite(chronologicalAge.Value))
-            return null;
-
-        var difference = crowdAge - chronologicalAge.Value;
-        if (!double.IsFinite(difference))
-            return null;
-
-        if (Math.Abs(difference) < 0.05)
-            return "about the same age as their chronological age";
-
-        var years = Math.Abs(difference).ToString("0.#", CultureInfo.InvariantCulture);
-        return difference < 0
-            ? $"{years} years below chronological age"
-            : $"{years} years above chronological age";
-    }
-
-    private static string BuildAgeImprovementTop10Line(
-        string athleteName,
-        string clock,
-        int place,
-        int? previousPlace,
-        string? previousAthlete,
-        double improvement,
-        string athleteUrl)
-    {
-        var placeText = CrowdOrdinal(place);
-        var movement = previousPlace.HasValue
-            ? previousPlace.Value > place ? $"climbed from {CrowdOrdinal(previousPlace.Value)} to {placeText}" : $"moved from {CrowdOrdinal(previousPlace.Value)} to {placeText}"
-            : place == 1 ? $"took {placeText} place" : $"entered the top 10 at {placeText}";
-        var leaderboardName = string.Equals(clock, "bortz", StringComparison.OrdinalIgnoreCase)
-            ? "Bortz Improvement"
-            : "Pheno Improvement";
-        var improvementText = FormatSignedYears(improvement);
-
-        var lead = !string.IsNullOrWhiteSpace(previousAthlete)
-            ? $"{athleteName} {movement} in the {leaderboardName} leaderboard, ahead of {previousAthlete}."
-            : $"{athleteName} {movement} in the {leaderboardName} leaderboard.";
-
-        return $"{lead}\n\nImprovement: {improvementText} years from worst to latest eligible result.\n\n{athleteUrl}";
-    }
-
-    private static string CrowdOrdinal(int n)
-    {
-        var suffix = (n % 100) is 11 or 12 or 13
-            ? "th"
-            : (n % 10) switch
-            {
-                1 => "st",
-                2 => "nd",
-                3 => "rd",
-                _ => "th"
-            };
-        return $"{n}{suffix}";
     }
 
     private static string BuildTop3LeaderboardIntro(XPostPhase? phase, string leagueDisplayName)
